Implement generation of assignment operators
Also add the --suppress-assignment option. We need to generate these
operators into the .cxx file instead of relying on the implicit inline
ones because in the file-per-type mode types of members are only forward-
declared in the header. Also, we don't want to assign the DOMDocument
member used to store the wildcard content.